refactor: useHighlightEventListener 훅 생성
