+++ frontend/src/components/highlight/components/HighlightEditor/index.tsx
@@ -1,15 +1,9 @@
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 
 import GrayHighlighterIcon from '@/assets/grayHighlighter.svg';
 import PrimaryHighlighterIcon from '@/assets/primaryHighlighter.svg';
-import {
-  EDITOR_ANSWER_CLASS_NAME,
-  EDITOR_LINE_CLASS_NAME,
-  HIGHLIGHT__TOGGLE_BUTTON_CLASS_NAME,
-  HIGHLIGHT_REMOVER_CLASS_NAME,
-} from '@/constants';
+import { EDITOR_ANSWER_CLASS_NAME, EDITOR_LINE_CLASS_NAME } from '@/constants';
 import { ReviewAnswerResponseData } from '@/types';
-import { findSelectionInfo, isTouchDevice } from '@/utils';
 
 import DragHighlightButtonContainer from '../DragHighlightButtonContainer';
 import EditorLineBlock from '../EditorLineBlock';
@@ -23,6 +17,7 @@ import {
   useLongPressHighlightButtonPosition,
   useLongPress,
   useEditableState,
+  useHighlightEventListener,
 } from './hooks';
 import * as S from './style';
 
@@ -81,63 +76,13 @@ const HighlightEditor = ({ questionId, answerList, handleErrorModal }: Highlight
 
   const { startPressTimer, clearPressTimer } = useLongPress({ handleLongPress: handleLongPressLine });
 
-  const hideHighlightButton = (e: MouseEvent | TouchEvent) => {
-    if (!isEditable) return;
-
-    const isInButton = (e.target as HTMLElement).closest(`.${HIGHLIGHT__TOGGLE_BUTTON_CLASS_NAME}`);
-    const isNotHighlightRemover = (e.target as HTMLElement).closest(`.${HIGHLIGHT_REMOVER_CLASS_NAME}`);
-
-    if (!isInButton) hideDragHighlightButton();
-    if (!isNotHighlightRemover) hideLongPressHighlightButton();
-  };
-
-  const showHighlightButton = () => {
-    if (!isEditable) return;
-    const selectionInfo = findSelectionInfo();
-    if (!selectionInfo) return;
-
-    const isAddingHighlight = checkHighlight(selectionInfo);
-    updateDragHighlightButtonPosition({ selectionInfo, isAddingHighlight });
-  };
-
-  /**
-   * document에 형광펜 이벤트 적용
-   */
-  const addHighlightEvent = () => {
-    document.addEventListener('mousedown', hideHighlightButton);
-    document.addEventListener('mouseup', showHighlightButton);
-    // NOTE: 터치가 가능한 기기에서는 touchstart, touchend 보다 selectionchange를 사용하는 게 오류가 없음
-    if (isTouchDevice()) {
-      document.addEventListener('selectionchange', showHighlightButton);
-      document.addEventListener('contextmenu', hideContextMenuInTouch);
-    }
-  };
-  /**
-   * 터치 브라우저에서, 글자 길게 선택 시 나오는 브라우저 기본 컨텍스트 메뉴 보이지 않게 처리하는 핸들러
-   * @param event
-   */
-  const hideContextMenuInTouch = (event: MouseEvent) => {
-    event.preventDefault();
-  };
-  /**
-   * document에 형광펜 이벤트 삭제
-   */
-  const removeHighlightEvent = () => {
-    document.removeEventListener('mouseup', showHighlightButton);
-    document.removeEventListener('mousedown', hideHighlightButton);
-    if (isTouchDevice()) {
-      document.removeEventListener('contextmenu', hideContextMenuInTouch);
-      document.removeEventListener('selectionChange', showHighlightButton);
-    }
-  };
-
-  useEffect(() => {
-    isEditable ? addHighlightEvent() : removeHighlightEvent();
-
-    return () => {
-      removeHighlightEvent();
-    };
-  }, [isEditable]);
+  useHighlightEventListener({
+    isEditable,
+    updateDragHighlightButtonPosition,
+    hideDragHighlightButton,
+    hideLongPressHighlightButton,
+    checkHighlight,
+  });
 
   return (
     <S.HighlightEditor ref={editorRef}>
